better logging on lambda
