+++ packages/lambda/src/functions/renderer.ts
@@ -21,6 +21,7 @@ import {getBrowserInstance} from './helpers/get-browser-instance';
 import {executablePath} from './helpers/get-chromium-executable-path';
 import {getCurrentRegionInFunction} from './helpers/get-current-region';
 import {lambdaWriteFile} from './helpers/io';
+import {onDownloadsHelper} from './helpers/on-downloads-logger';
 import {
 	getTmpDirStateIfENoSp,
 	writeLambdaError,
@@ -111,8 +112,6 @@ const renderHandler = async (
 		)}`
 	);
 
-	const downloads: Record<string, number> = {};
-
 	const resolvedProps = await resolvedPropsPromise;
 	const serializedInputPropsWithCustomSchema = await inputPropsPromise;
 
@@ -185,36 +184,7 @@ const renderHandler = async (
 			crf: params.crf ?? null,
 			pixelFormat: params.pixelFormat ?? RenderInternals.DEFAULT_PIXEL_FORMAT,
 			proResProfile: params.proResProfile,
-			onDownload: (src: string) => {
-				console.log('Downloading', src);
-				downloads[src] = 0;
-				return ({percent, downloaded}) => {
-					if (percent === null) {
-						console.log(
-							`Download progress (${src}): ${downloaded} bytes. Don't know final size of download, no Content-Length header.`
-						);
-						return;
-					}
-
-					if (
-						// Only report every 10% change
-						downloads[src] > percent - 0.1 &&
-						percent !== 1
-					) {
-						return;
-					}
-
-					downloads[src] = percent;
-					console.log(
-						`Download progress (${src}): ${downloaded} bytes, ${(
-							percent * 100
-						).toFixed(1)}%`
-					);
-					if (percent === 1) {
-						console.log(`Download complete: ${src}`);
-					}
-				};
-			},
+			onDownload: onDownloadsHelper(),
 			overwrite: false,
 			chromiumOptions: params.chromiumOptions,
 			scale: params.scale,
